+++ berkeley-function-call-leaderboard/bfcl/utils.py
@@ -4,7 +4,7 @@ import re
 from pathlib import Path
 from typing import Union
 
-from bfcl.constant import VERSION_PREFIX
+from bfcl.constant import TEST_COLLECTION_MAPPING, TEST_FILE_MAPPING, VERSION_PREFIX
 
 
 def extract_test_category(input_string: Union[str, Path]) -> str:
@@ -19,6 +19,10 @@ def extract_test_category(input_string: Union[str, Path]) -> str:
         raise ValueError(
             f"Could not extract the test category from the input string: {input_string}"
         )
+
+
+def extract_test_category_from_id(test_entry_id: str) -> str:
+    return test_entry_id.rsplit("_", 1)[0]
 
 
 def find_file_with_suffix(folder_path: Path, suffix: str) -> Path:
@@ -64,12 +68,15 @@ def is_sql(test_category):
     return "sql" in test_category
 
 
-def load_file(file_path):
+def load_file(file_path, sort_by_id=False):
     result = []
     with open(file_path) as f:
         file = f.readlines()
         for line in file:
             result.append(json.loads(line))
+
+    if sort_by_id:
+        result.sort(key=sort_key)
     return result
 
 
@@ -108,14 +115,48 @@ def make_json_serializable(value):
             return str(value)
 
 
+def sort_key(entry):
+    """
+    Index comes in two forms: TestCategory_Index or TestCategory_Index-FuncDocSubIndex-PromptSubIndex; both 0-indexed.
+
+    TestCategory_Index: For example, `simple_20` means the 21st entry in the `simple` test category.
+
+    TestCategory_Index-FuncDocSubIndex-PromptSubIndex is used when there are multiple prompts for a single function doc; this only happens in the live dataset.
+    FuncDocSubIndex increments for each unique function doc.
+    PromptSubIndex is per function doc. It resets to 0 for each function doc.
+        For example, `live_simple_19-3-15` means the 20th entry in the `live_simple` test category.
+        This entry has the 4th unique function doc and the 16th prompt for that function doc (there are at least 15 other prompts for this same function doc in this category).
+
+    In either case, the universal index is enough to sort the entries.
+    """
+    parts = entry["id"].rsplit("_", 1)
+    test_category, index = parts[0], parts[1]
+    # This handles the case where the index is in the form TestCategory_Index-FuncDocSubIndex-PromptSubIndex
+    if "-" in index:
+        index = index.split("-")[0]
+    return (test_category, int(index))
+
+
 def is_function_calling_format_output(decoded_output):
-    # Ensure the output is a list of dictionaries
-    if type(decoded_output) == list:
-        for item in decoded_output:
-            if type(item) != dict:
-                return False
-        return True
-    return False
+    """
+    Ensure the output is a list of dictionaries of the form:
+    `[{func1: {param1: val1, param2: val2, ...}}, {func2: {param1: val1, param2: val2, ...}}, ...]`
+    Sometimes the model handler's `decode_ast` method will return successfully, but the output is not in the correct format, and that will mess up the downstream evaluation that expects this format.
+    This is especially the case when the model doesn't predict any function calls, and the output is an human-readable string.
+    Note: Empty list `[]` is considered the correct format in this check.
+    """
+    if type(decoded_output) != list:
+        return False
+    for item in decoded_output:
+        if type(item) != dict:
+            return False
+        # Check for `{func1: {param1: val1, param2: val2, ...}}`, should only have one key-value pair
+        if len(item) != 1:
+            return False
+        # Check for `{param1: val1, param2: val2, ...}`; the parameter-value pairs should be a dictionary
+        if type(list(item.values())[0]) != dict:
+            return False
+    return True
 
 
 def is_executable_format_output(decoded_output):
@@ -149,3 +190,33 @@ def is_empty_output(decoded_output):
     if len(decoded_output) == 1 and len(decoded_output[0]) == 0:
         return True
     return False
+
+
+def check_api_key_supplied() -> bool:
+    """
+    This function checks if the four API Keys needed for the executable categoreis are provided. If not, those categories will be skipped.
+    """
+    ENV_VARS = ("GEOCODE_API_KEY", "RAPID_API_KEY", "OMDB_API_KEY", "EXCHANGERATE_API_KEY")
+    for var in ENV_VARS:
+        if os.getenv(var) == "":
+            return False
+    return True
+
+
+def parse_test_category_argument(test_category_args):
+    test_name_total = set()
+    test_filename_total = set()
+
+    for test_category in test_category_args:
+        if test_category in TEST_COLLECTION_MAPPING:
+            for test_name in TEST_COLLECTION_MAPPING[test_category]:
+                test_name_total.add(test_name)
+                test_filename_total.add(TEST_FILE_MAPPING[test_name])
+        elif test_category in TEST_FILE_MAPPING:
+            test_name_total.add(test_category)
+            test_filename_total.add(TEST_FILE_MAPPING[test_category])
+        else:
+            # Invalid test category name
+            raise Exception(f"Invalid test category name provided: {test_category}")
+
+    return sorted(list(test_filename_total)), sorted(list(test_name_total))